# docs/api-documentation.yaml
openapi: 3.0.3
info:
  title: Voke E-commerce API
  description: |
    Complete API for the Voke electronic store.
    
    This API provides endpoints to manage products, categories, shopping cart, and user authentication.
    
    ## Main Features:
    - Complete electronic product management
    - Advanced filtering by categories, prices, and brands
    - Product search functionality
    - Result pagination
    - Featured products and offers
    - Shopping cart with session support
    - JWT user authentication
    - Interactive Swagger documentation
    
    ## Authentication
    The API uses JWT (JSON Web Tokens) for authentication. For protected endpoints,
    include the token in the Authorization header:
    ```
    Authorization: Bearer <your_jwt_token>
    ```
    
    ## Shopping Cart
    The cart supports both anonymous users (using session_id) and authenticated
    users (using user_id). When logging in, the cart is automatically migrated.
    
    ## Base URL
    - **Development:** `http://localhost:3001/api`
    - **Production:** `https://tu-app-voke-backend-7da6ed58e5fc.herokuapp.com/api`
  version: 1.0.0
  contact:
    name: Voke Team
    email: support@voke.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://tu-app-voke-backend-7da6ed58e5fc.herokuapp.com/api
    description: Production Server (Heroku)
  - url: http://localhost:3001/api
    description: Development Server
paths:
...
  /auth/register:
    post:
      summary: Register new user
      description: Creates a new user account en el sistema
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: Unique user email
                  example: usuario@example.com
                password:
                  type: string
                  minLength: 6
                  description: User password (minimum 6 characters)
                  example: mi_password_123
                first_name:
                  type: string
                  minLength: 2
                  description: User first name
                  example: Juan
                last_name:
                  type: string
                  description: User last name (optional)
                  example: Pérez
                phone:
                  type: string
                  description: User phone (optional)
                  example: +1234567890
                session_id:
                  type: string
                  description: Session ID para migrar carrito (optional)
                  example: session_abc123
              required:
                - email
                - password
                - first_name
      responses:
        '201':
          description: User registered successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: User registered successfully
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        description: JWT token para autenticación
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '409':
          description: Email already registered
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/login:
    post:
      summary: Login user
      description: Authenticates a user y devuelve un token JWT
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                email:
                  type: string
                  format: email
                  description: User email
                  example: usuario@example.com
                password:
                  type: string
                  description: User password
                  example: mi_password_123
                session_id:
                  type: string
                  description: Session ID para migrar carrito (optional)
                  example: session_abc123
              required:
                - email
                - password
      responses:
        '200':
          description: Login successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Login successful
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
                      token:
                        type: string
                        description: JWT token para autenticación
                        example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
        '401':
          description: Invalid credentials
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '429':
          description: Too many login attempts
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/logout:
    post:
      summary: Logout user
      description: Logs out the user autenticado
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          description: JWT token del usuario
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Logout successful
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Logout successful
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/me:
    get:
      summary: Get user profile
      description: Gets profile information del usuario autenticado
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          description: JWT token del usuario
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      responses:
        '200':
          description: Profile retrieved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/profile:
    put:
      summary: Update user profile
      description: Updates profile information del usuario autenticado
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          description: JWT token del usuario
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                first_name:
                  type: string
                  minLength: 2
                  description: New nombre del usuario
                  example: Juan Carlos
                last_name:
                  type: string
                  description: New apellido del usuario
                  example: Pérez García
                phone:
                  type: string
                  description: New teléfono del usuario
                  example: +1234567890
      responses:
        '200':
          description: Profile updated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Profile updated successfully
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Invalid data
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/change-password:
    put:
      summary: Change password
      description: Changes user password autenticado
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          description: JWT token del usuario
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                current_password:
                  type: string
                  description: Current password del usuario
                  example: mi_password_actual
                new_password:
                  type: string
                  minLength: 6
                  description: New password (minimum 6 characters)
                  example: mi_nueva_password_123
              required:
                - current_password
                - new_password
      responses:
        '200':
          description: Password changed successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Password changed successfully
        '400':
          description: Invalid data o contraseña actual incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/verify-token:
    post:
      summary: Verify JWT token
      description: Verifies if a JWT token is valid y devuelve información del usuario
      tags:
        - Authentication
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                token:
                  type: string
                  description: JWT token a verificar
                  example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
              required:
                - token
      responses:
        '200':
          description: Token valid
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Token valid
                  data:
                    type: object
                    properties:
                      user:
                        $ref: '#/components/schemas/User'
        '400':
          description: Token required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid or expired token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /auth/account:
    delete:
      summary: Deactivate account
      description: Deactivates user account autenticado (soft delete)
      tags:
        - Authentication
      parameters:
        - name: Authorization
          in: header
          description: JWT token del usuario
          required: true
          schema:
            type: string
            example: Bearer eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  description: User password para confirmar la desactivación
                  example: mi_password_123
              required:
                - password
      responses:
        '200':
          description: Account deactivated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: Account deactivated successfully
        '400':
          description: Password required
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '401':
          description: Invalid token o contraseña incorrecta
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Internal server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          format: uuid
          description: Unique user ID
          example: user123e4567-e89b-12d3-a456-426614174000
        email:
          type: string
          format: email
          description: User email
          example: usuario@example.com
        first_name:
          type: string
          description: User first name
          example: Juan
        last_name:
          type: string
          description: User last name
          example: Pérez
        phone:
          type: string
          description: User phone
          example: +1234567890
        is_active:
          type: boolean
          description: If the user is active
          example: true
        created_at:
          type: string
          format: date-time
          description: Creation date
          example: 2025-08-22T16:33:53.03547+00:00
        updated_at:
          type: string
          format: date-time
          description: Last update date
          example: 2025-08-22T16:33:53.03547+00:00
      required:
        - id
        - email
        - first_name
        - is_active
...
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        message:
          type: string
          description: Mensaje de error
          example: Internal server error
        error:
          type: string
          description: Detalles del error
          example: Database connection failed
      required:
        - success
        - message
...